+++ src/REvoDesign/shortcuts/tools/esm2.py
@@ -10,12 +10,12 @@ import string
 from typing import List, Literal, Optional, Tuple
 
 import pandas as pd
-import torch
+
 from Bio import SeqIO
 from immutabledict import immutabledict
 from tqdm import tqdm
 
-from REvoDesign.basic import ThirdPartyModuleAbstract
+from REvoDesign.basic import ThirdPartyModuleAbstract,TorchModuleAbstract
 from REvoDesign.bootstrap.set_config import is_package_installed
 from REvoDesign.tools.utils import get_cited, require_installed
 
@@ -80,7 +80,7 @@ def label_row(row, sequence, token_probs, alphabet, offset_idx):
 
 
 @require_installed
-class Esm1v(ThirdPartyModuleAbstract):
+class Esm1v(ThirdPartyModuleAbstract,TorchModuleAbstract):
     name: str = "esm1v"
     installed: bool = is_package_installed('esm2')
 
@@ -149,9 +149,13 @@ class Esm1v(ThirdPartyModuleAbstract):
                     for idx, mut in itertools.product(range(0, len(self.sequence)), alphabet)
                 ], columns=[self.mutation_col])
         return df_dms
+    
+            
+    
 
     @get_cited
     def predict(self):
+        import torch
         from esm2 import MSATransformer, pretrained  # type: ignore
 
         # Load the deep mutational scan
@@ -159,9 +163,10 @@ class Esm1v(ThirdPartyModuleAbstract):
 
         # inference for each model
         for model_name in self.model_names:
-            if self.checkpoint_dir is not None and os.path.isdir(model_name):
-                model_path = os.path.join(self.checkpoint_dir, f'{model_name}.pt')
+            if self.checkpoint_dir is not None and os.path.isdir(self.checkpoint_dir) and  os.path.isfile(model_path := (os.path.join(self.checkpoint_dir, f'{model_name}.pt'))):
+                print(f"Loading model from {model_path}")
             else:
+                print(f'Fetching model {model_name}')
                 model_path = model_name
             model, alphabet = pretrained.load_model_and_alphabet(
                 model_path if os.path.isfile(model_path) else os.path.basename(model_name).rstrip('.pt'))
@@ -251,6 +256,8 @@ class Esm1v(ThirdPartyModuleAbstract):
         df.to_csv(self.dms_output)
 
     def compute_pppl(self, row, sequence, model, alphabet, offset_idx):
+        import torch
+
         wt, idx, mt = row[0], int(row[1:-1]) - offset_idx, row[-1]
         assert sequence[idx] == wt, "The listed wildtype does not match the provided sequence"
 
@@ -278,6 +285,9 @@ class Esm1v(ThirdPartyModuleAbstract):
             log_probs.append(token_probs[0, i, alphabet.get_idx(sequence[i])].item())  # vocab size
         return sum(log_probs)
 
+
+ 
+        
     __bibtex__ = {
         'ESM': """@article{rives2019biological,
   author={Rives, Alexander and Meier, Joshua and Sercu, Tom and Goyal, Siddharth and Lin, Zeming and Liu, Jason and Guo, Demi and Ott, Myle and Zitnick, C. Lawrence and Ma, Jerry and Fergus, Rob},
@@ -341,3 +351,6 @@ def shortcut_esm1v(
         device=device,
     )
     predictor.predict()
+    predictor.cleanup()
+
+    del predictor